re-integrate gateway client into purplet
