+++ packages/purplet/src/lib/env.ts
@@ -3,6 +3,7 @@ import structuredClone from '@ungap/structured-clone';
 import fetch, { Request, Response } from 'node-fetch';
 import path from 'path';
 import { config } from 'dotenv';
+import { WebSocket } from 'ws';
 
 const envFile = path.resolve(process.cwd(), '.env');
 
@@ -12,6 +13,7 @@ export function setupEnv(isDev: boolean) {
   (global as any).fetch ??= fetch;
   (global as any).Request ??= Request;
   (global as any).Response ??= Response;
+  (global as any).WebSocket ??= WebSocket;
 }
 
 export function getEnvVar(name: string): string | null {
