+++ packages/purplet/src/structures/guild.ts
@@ -24,22 +24,3 @@ export type EmptyGuild = PartialClass<
   'id' | 'fetch'
 >;
 export const EmptyGuild = createPartialClass<EmptyGuild>(Guild);
-
-const restclient = new REST({ }).setToken("token") //id prefer if token was just a property passed to constructor
-
-restclient.post('/some url', {
-  body: {
-    json_body: true
-  },
-  query: new URLSearchParams(),
-  headers: {
-    // headers
-  },
-  files: [
-    {
-      data: new Uint8Array(), // i guess allow string, arraybuffer, or whatever you want.
-      name: 'filename'
-    }
-  ],
-  reason: '', // this is translated to X-Audit-Log-Reason header
-})
